Rebuild the Project and add check to SceneManager
This commit rebuilds the project so it is upto date
along with commenting out the setting of the start
screen within the Game1 class. Since the update and
draw methods within the scene manager expect a screen
to always be there this would cause an error and we
should allow for the screen to be null if we want it
to be. For this reason a null check has been added before
updating/drawing.