Rework keyboard input
The handler now sends to as single method in the subscriber
the event args contains a dict with the keystate as its
key and the keys as the values in said state.

There was previously a bug that told the
subscribers that a key was being repressd/ released
when key was not. This bug has now been fixed.