Converts the EventManager from a singleton
The event manager which used to be a singleton
meant that each screen had to share the same
EventManager which is not great when you want
to have multiple screens loaded (but not updating)
at the same time. With this refactor the each screen
now has its own event manager which is created within
the GameScreen class so it is accessable to the screens
if needed. The Created event manager is injected into the
AssetManager which injects it into the AIfactory so it can
still register all events for the Ai