Add overloads to CreateAsset without minds
This commit adds 4 new overloads to the CreateAsset mind
these overloads are esentialy the same as the other ones
except they dont take a second generic paramater for the mind
this allows for assets to be created without minds.

This also fixes a bug where the LevelLoader was not returning the
current _uid of the tiles which was causing conflicts when in use.