Change the CollisionManagers update paramaters
I had just modeled the colllision manager to take the
assets and their minds as a dictonary of tuples however
it is not the responsibilty of the asset manager to create
that dict for the collision manager. It now instead just
takes the asset and ai dicts.

To reflect this the RemoveNonCollidableAssets has now been
renamed to GetCollidableAssets which takes the two dicts
and converts them into a dict of tuples for convinence