Further optimise collisions
The quadtree now has a method which will
return a list of lists that will be the collidable
assets in each quadrant.

The collision manager will now call this method and
loop through each quadrant of assets. when looping through
each asset within the quadrant and checking it against every other
we record what has been checked against each other so no duplicate
checks are made.

This 1 - halfs the number of checks we need to do and reduces the
amount of calls we need to make to the quad tree which would have had
to do multiple recursive searches per update now it only has to return
A list of quadrant collidables